refactor: shared logs utils
