+++ shared/package.json
@@ -22,6 +22,7 @@
     "@open-policy-agent/opa-wasm": "1.8.0",
     "@reduxjs/toolkit": "1.9.0",
     "electron": "^21.2.3",
+    "electron-log": "^4.4.8",
     "electron-store": "^8.1.0",
     "lodash": "4.17.21",
     "nodemon": "2.0.20",
