+++ shared/src/utils/index.ts
@@ -2,5 +2,6 @@ export * from './electronStore';
 export * from './env';
 export * from './fileSystem';
 export * from './hotkey';
+export * from './logs';
 export * from './selectors';
 export * from './shell';
